+++ webapi/src/main/scala/org/knora/webapi/instrumentation/prometheus/PrometheusApp.scala
@@ -1,19 +1,17 @@
 package org.knora.webapi.instrumentation.prometheus
 
 import zhttp.http._
-import zhttp.service.Server
 import zio.ZIO
 import zio.metrics.connectors.prometheus.PrometheusPublisher
 
+/**
+ * Provides the '/metrics' endpoint serving the metrics in prometheus format.
+ */
 object PrometheusApp {
 
-  private lazy val prometheusRouter =
+  def apply(): HttpApp[PrometheusPublisher, Nothing] =
     Http
       .collectZIO[Request] { case Method.GET -> !! / "metrics" =>
         ZIO.serviceWithZIO[PrometheusPublisher](_.get.map(Response.text))
       }
-
-  val make =
-    Server.app(prometheusRouter)
-
 }
